Cleaned up the garbage, and put the folders in order
