--- VKApp_KonishchevIvan/VKApp_KonishchevIvan/Controllers/ViewCollectionCell/FullImageCell/FullImageCell.swift
@@ -1,15 +1,0 @@
-//
-//  FullImageCell.swift
-//  VKApp_KonishchevIvan
-//
-//  Created by Ivan Konishchev on 18.03.2022.
-//
-
-import UIKit
-
-class FullImageCell: UICollectionViewCell {
-    
-    @IBOutlet weak var fullimageCellSet: UIImageView!
-    
-    
-}
