+++ VKApp_KonishchevIvan/VKApp_KonishchevIvan/Model/AllProtocols/TableViewDelegate.swift
@@ -1,0 +1,13 @@
+//
+//  TableViewDelegate.swift
+//  VKApp_KonishchevIvan
+//
+//  Created by Ivan Konishchev on 01.04.2022.
+//
+
+import Foundation
+
+protocol TableViewDelegate {
+    var nextViewData: [ImageAndLikeData] {set get}
+    func selectRow(nextViewData: [ImageAndLikeData])
+}
