+++ VKApp_KonishchevIvan/VKApp_KonishchevIvan/Model/AllProtocols/UserGroupTableViewDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  UserGroupTableViewDelegate.swift
+//  VKApp_KonishchevIvan
+//
+//  Created by Ivan Konishchev on 01.04.2022.
+//
+
+import Foundation
+
+protocol UserGroupTableViewDelegate {
+    func addNewGroup (_ group: [AllUserGroups], allGroup: [AllUserGroups] )
+}
