--- VKApp_KonishchevIvan/VKApp_KonishchevIvan/Model/ProtocolLikeDelegate.swift
@@ -1,14 +1,0 @@
-//
-//  ProtocolLikeDelegate.swift
-//  VKApp_KonishchevIvan
-//
-//  Created by Ivan Konishchev on 26.03.2022.
-//
-
-import UIKit
-
- protocol ProtocolLikeDelegate: AnyObject {
-    func getCountLike(for indexPath: IndexPath) -> [Int: Bool]
-    func setCountLike(countLike: Int, likeStatus: Bool, for indexPath: IndexPath)
-  //  func getCountLikeGallary(for numberImage: Int) ->[Int: Bool]
-}
