+++ VKApp_KonishchevIvan/VKApp_KonishchevIvan/ViewControllers/NewGroupTableViewController.swift
@@ -8,8 +8,6 @@
 import UIKit
 
 class NewGroupTableViewController: UITableViewController, UISearchBarDelegate {
-
-
     
     @IBOutlet weak var searchBar: CustomCodeSearchBar!
